refactor: isolate daemon loader and builtin dispatch adapters
